Refactor: Extract native/shared/each.js for server/client coordination
`getItemID`, `getEachData`, item-marker prefix, and key encode/decode
were duplicated between `server.js` and `blocks/each.js` with a
keep-in-sync comment as the only enforcement. The wire format that
bridges SSR and client hydration is now sourced from one module.

- New `engines/native/shared/each.js` exporting SUI_ITEM_MARKER,
  getItemID, getEachData, encodeItemKey, decodeItemKey
- `server.js` and `blocks/each.js` import from shared
- `blocks/each.js` also imports BLOCK_MARKER + BLOCK_CLOSE_PREFIX
  for its server-group walker (was inlining literals)
